user.java changed fieldtype from  latitude and longitude integer-->double

+++ src/main/java/ch/uzh/ifi/hase/soprafs21/entity/User.java
@@ -45,10 +45,10 @@ public class User implements Serializable {
     private Timestamp timestamp;
 
     @Column
-    private Integer longitude;
+    private Double longitude;
 
     @Column
-    private Integer latitude;
+    private Double latitude;
 
 
 
@@ -56,19 +56,19 @@ public class User implements Serializable {
         return serialVersionUID;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
